- sets the base url for migration
